+++ CHANGES.rst
@@ -19,6 +19,7 @@ Bug fixes:
 
 API relevant changes:
 
+* `BufferProcessor` uses `data` instead of `buffer` for data storage (#292)
 * `DBNBeatTrackingProcessor` expects 1D inputs (#299)
 
 Other changes: